Generica YAML component parsing
Implemented code that will parse a YAML file containing FHIR components and create the appropriate FHIR resource. Currently it does not insert the components anywhere as where the components go still needs to be decided. There is currently no examples I have a a valueTime so I have decided to ignore it( I believe an error would be inappropriate as this can be done I just need to see how its formatted). I will message Dr.Dolin about this and try to get some examples of valueTime usage in the wild.  The push will look off because this push also introduces the improvements you made to the code. as well as files that were not present before.